Optimized mc_network_lookup for repeated calls.
Now, it holds the hdr, the map (which is the network map volume) in persistent memory.

Additionally, it stores as a string the last path to the mapfile.

If this is undefined (i.e. it is the first time running mc_network_lookup)
it will run as usual. Similarly, if the current value of mapFile is different
than the one used at last run, it will run as usual.

However, if mapFile matches oldmapFile, it will not bother loading hdr or map
and instead use the existing values, which should hugely increase speed