Fixed ktaub to properly check for matlab version prior to running.
MATLAB versions appear to consist of four digits. In the past, it first checked
if your first digit is 7 or greater.

Next, it checks if your second digit is 9 or greater.
The motivation was to check if the version is greater than 7.9

However, this logic check would also fail for version 8.0, etc. Instead,
if you have version 7, we care about the second digit. If you have greater
than version 7, we don't care about the second digit.

Added logic check to remedy this