Big bug detected (and hopefully now corrected).
netmask is a cell array of logicals that allows you to quickly
subset your flattened connectomes to look at just one network cell/intersection.

However, netmask's indexing was off, so netmask(1,1) corresponded to the
intersection of network 1 and 1. This doesn't seem so bad, but IT IS!

Since cell arrays in matlab are 1-indexed, there is no place to put the
intersection of network 0-0.

The internal TakGraph code fixes this problem by first translating
the networks to something called netSort. It does this with
sort(unique(nets)) so that you get a sorted list of unique networks.

It then does 1-indexing into this vector. So, anywhere it stores something
at, say, (2,3), it means the intersection of the SECOND network from netSort
with the THIRD network from netSort. If you are using the typical network labels,
this will be network 1 with network 2 (because there is a network 0 which is
the first network).

We now doing something very similar for netmask.