Clean up SOM
This is a little confusing. At some point in the development of
ConnTool code (that has since been merged to public), the "som" directory
was reintroduced. In order to quash this, I just tacked on a commit
that manually deleted all of the SOM folders.

At some point, @heffjos reverted this commit, which was apparently necessary in order for to deal with a merge conflict or something. This had the net
effect of reintroducing SOM to the FirstLevel_alpha branch (which also contains public).

As we prepare to merge FirstLevel_alpha into FirstLevel_beta, in addition to
bringing public along with it, it was going to also reintroduced SOM.

So, I am reverting Joe's reversion. This effectively deletes all of the SOM stuff.

I've checked, and this will cleanly merge into FirstLevel_beta, and eventually into public, with this reversion included.

Revert "Revert "Removed 'som' that was accidently there for Robert's convience of testing and simultaneously updating.""

This reverts commit 86125c5f125d8335010b99bdfcc9baf794c8749d.